backend/pyproject.toml
[project]
name = "hr-avatar-backend"
version = "1.0.0"
description = "FastAPI backend for HR Avatar interview system with state machine"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
    "sqlalchemy>=2.0.23",
    "alembic>=1.12.1",
    "redis>=5.0.1",
    "pydantic[email]>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "openai>=1.3.0",
    "openai-whisper>=20231117",
    "numpy>=1.26.0",
    "psycopg2-binary>=2.9.7",
    "python-jose[cryptography]>=3.3.0",
    "python-dotenv>=1.0.0",
    "email-validator>=2.1.0",
    "aiofiles>=23.2.1",
    "librosa>=0.10.1",
    "soundfile>=0.12.1",
    "passlib>=1.7.4",
    "bcrypt>=4.3.0",
    "elevenlabs>=0.2.26",
    "pydub>=0.25.1",
    "aiohttp>=3.12.15",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0"
]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0"
]
